// src/app/core/interfaces/sidebar.ts
export interface Sidebar {}

// Generated by https://quicktype.io

export interface DataLocalStorage {
	mensaje_return: MensajeReturn;
	token: string;
	result: Result;
	data_menu: DataMenu;
	data_company: DataCompany;
}
export interface ReturnMenu {
	id?: number;
	txt?: string;
	route?: string;
	menu_name?: string;
	icon?: string;
}
// Generated by https://quicktype.io

export interface Data {
	mensaje_return: MensajeReturn;
	token: string;
	result: Result;
	data_menu: DataMenu;
	data_company: DataCompany;
}

export interface DataCompany {
	id_user_company: number;
	id_user: number;
	id_language: number;
	id_user_type: number;
	user_type: string;
	status_user: number;
	accept_term_condition: number;
	id_registration_method: number;
	first_name: string;
	surname: string;
	alias_username: string;
	email: string;
	email_validate_code: string;
	validate_mail: number;
	email_valid_until: string;
	email_validated_on: null;
	token: string;
	password: string;
	change_password: number;
	connection_status: number;
	ip_last_connection: string;
	conexion_date: string;
	registration_date: string;
	id_company: number;
	id_courier: number;
	address: string;
	company_name: string;
	img_logo: string;
	id_company_type: number;
	company_type: string;
}

export interface DataMenu {
	return_menu: ReturnMenu[];
	return_profiles: ReturnProfile[];
}

export interface ReturnMenu {
	id?: number;
	txt?: string;
	route?: string;
	menu_name?: string;
	icon?: string;
}

export interface ReturnProfile {
	txt: string;
	route: string;
	icon: string;
}

export interface MensajeReturn {
	ERROR: boolean;
	ERROR_MENSAGGE: string;
	CODE: number;
}

export interface Result {
	id_user: number;
	first_name: string;
	surname: string;
	alias_username: string;
	email: string;
	validate_mail: number;
	conexion_date: string;
	id_language: number;
	acronym_language: string;
	language_name: string;
	icon: null;
	id_user_type: number;
	user_type: string;
}

// NAVBAR MENU
export interface NavbarMenu {
	txt: string;
	route: string;
}
// NAVBAR DROPDOWN
export interface DropdownMenu {
	txt: string;
	route: string;
	icon: string;
}
export interface NavbarMenuIcon {
	txt: string;
	route: string;
	icon: string;
}

export interface NotificationMenu {
	id_topic_audit?: number;
	locker_name?: string;
	order_number?: string;
	registration_date?: Date;
	time_elapsed?: number;
	topic_name?: string | number;
	tracking_number?: string;
}

export interface UsersMenu {
	company_employee_id?: string;
	description_audit?: string;
	full_name?: string;
	id_topic_audit?: number;
	locker_name?: string;
	registration_date?: string;
	time_elapsed?: number;
	topic_name?: string;
}